backend_python/pyproject.toml
[project]
name = "backend-python"
version = "2.0.3"
requires-python = ">=3.12"
dependencies = [
    "email-validator>=2.3.0",
    "fastapi>=0.128.0",
    "firebase-admin>=7.1.0",
    "google-cloud-aiplatform>=1.133.0",
    "google-cloud-firestore>=2.23.0",
    "google-genai>=1.59.0",
    "google-generativeai>=0.8.6",
    "httpx>=0.28.1",
    "langchain>=1.2.6",
    "langchain-core>=1.2.7",
    "langchain-google-genai>=4.2.0",
    "langchain-google-vertexai>=3.2.1",
    "langgraph>=1.0.6",
    "pillow>=12.1.0",
    "pinecone-client>=6.0.0",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.18",
    "uvicorn[standard]>=0.40.0",
    "pydantic-settings>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
]

[tool.pytest.ini_options]
minversion = "8.0"
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "integration: marks tests as integration tests (requiring live API)",
]

[tool.coverage.run]
source = ["src"]
omit = ["*/tests/*", "*/__pycache__/*", "*/.venv/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]
